+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/past/checker/MethodSignature.java
@@ -1,8 +1,10 @@
 package org.openprovenance.prov.template.compiler.past.checker;
 
+import org.openprovenance.prov.template.compiler.past.annotations.PastAnnotation;
 import org.openprovenance.prov.template.compiler.past.type.TypeName;
 
 import javax.lang.model.element.Modifier;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -14,6 +16,7 @@ public class MethodSignature {
     public final List<TypeName> typeVariables;
     public final boolean isStatic;
     public final Set<Modifier> modifiers;
+    public final List<PastAnnotation> annotations=new ArrayList<>();
 
     public MethodSignature(String name, List<TypeName> parameterTypes, TypeName returnType,
                            List<TypeName> typeVariables, boolean isStatic, Set<Modifier> modifiers) {
@@ -30,13 +33,20 @@ public class MethodSignature {
         this(name, parameterTypes, returnType, typeVariables, isStatic, Collections.emptySet());
     }
 
+    public List<PastAnnotation> getAnnotations() {
+        return annotations;
+    }
+
     @Override
     public String toString() {
         return "MethodSignature{" +
                 "name='" + name + '\'' +
                 ", parameterTypes=" + parameterTypes +
                 ", returnType=" + returnType +
+                ", typeVariables=" + typeVariables +
                 ", isStatic=" + isStatic +
+                ", modifiers=" + modifiers +
+                ", annotations=" + annotations +
                 '}';
     }
 }